% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_em_om.R
\name{adjust_emc_omc}
\alias{adjust_emc_omc}
\title{Adust for exposure misclassification and outcome misclassification.}
\usage{
adjust_emc_omc(
  data,
  exposure,
  outcome,
  confounders = NULL,
  x_model_coefs = NULL,
  y_model_coefs = NULL,
  x1y0_model_coefs = NULL,
  x0y1_model_coefs = NULL,
  x1y1_model_coefs = NULL,
  level = 0.95
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{adjust_emc_omc()} was renamed to \code{adjust_em_om()}
}
\keyword{internal}
