% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_em_sel.R
\name{adjust_emc_sel}
\alias{adjust_emc_sel}
\title{Adust for exposure misclassification and selection bias.}
\usage{
adjust_emc_sel(data_observed, x_model_coefs, s_model_coefs, level = 0.95)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{adjust_emc_sel()} was renamed to \code{adjust_em_sel()}
}
\keyword{internal}
