% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_multinom_uc_emc.R
\name{adjust_multinom_uc_emc}
\alias{adjust_multinom_uc_emc}
\title{Adust for uncontrolled confounding and exposure misclassification.}
\usage{
adjust_multinom_uc_emc(
  data,
  exposure,
  outcome,
  confounders = NULL,
  x1u0_model_coefs,
  x0u1_model_coefs,
  x1u1_model_coefs,
  level = 0.95
)
}
\arguments{
\item{data}{Dataframe for analysis.}

\item{exposure}{String name of the exposure variable.}

\item{outcome}{String name of the outcome variable.}

\item{confounders}{String name(s) of the confounder(s).
A maximum of three confounders are allowed.}

\item{x1u0_model_coefs}{The regression coefficients corresponding to the
model: \ifelse{html}{\out{log(P(X=1,U=0)/P(X=0,U=0)) =
&gamma;<sub>1,0</sub> + &gamma;<sub>1,1</sub>X* +
&gamma;<sub>1,2</sub>Y + &gamma;<sub>1,2+j</sub>C<sub>j</sub>, }
where X is the true (binary) exposure, U is the (binary) unmeasured
confounder, X* is the (binary) misclassified exposure, Y is the (binary)
outcome, C represents the vector of (binary) measured confounders (if any),
and j corresponds to the number of measured
confounders.}{\eqn{log(P(X=1,U=0)/P(X=0,U=0)) =}}}

\item{x0u1_model_coefs}{The regression coefficients corresponding to the
model: \ifelse{html}{\out{log(P(X=0,U=1)/P(X=0,U=0)) =
&gamma;<sub>2,0</sub> + &gamma;<sub>2,1</sub>X* +
&gamma;<sub>2,2</sub>Y + &gamma;<sub>2,2+j</sub>C<sub>j</sub>, }
where X is the true (binary) exposure, U is the (binary) unmeasured
confounder, X* is the (binary) misclassified exposure, Y is the (binary)
outcome, C represents the vector of (binary) measured confounders (if any),
and j corresponds to the number of measured
confounders.}{\eqn{log(P(X=0,U=1)/P(X=0,U=0)) =}}}

\item{x1u1_model_coefs}{The regression coefficients corresponding to the
model: \ifelse{html}{\out{log(P(X=1,U=1)/P(X=0,U=0)) =
&gamma;<sub>3,0</sub> + &gamma;<sub>3,1</sub>X* +
&gamma;<sub>3,2</sub>Y + &gamma;<sub>3,2+j</sub>C<sub>j</sub>, }
where X is the true (binary) exposure, U is the (binary) unmeasured
confounder, X* is the (binary) misclassified exposure, Y is the (binary)
outcome, C represents the vector of (binary) measured confounders (if any),
and j corresponds to the number of measured
confounders.}{\eqn{log(P(X=1,U=1)/P(X=0,U=0)) =}}}

\item{level}{Value from 0-1 representing the full range of the confidence
interval. Default is 0.95.}
}
\description{
\code{adjust_multinom_uc_emc} returns the exposure-outcome odds ratio and
confidence interval, adjusted for uncontrolled confounding and exposure
misclassificaiton.
}
\details{
This function uses one bias model, a multinomial logistic regression model,
to predict the uncontrolled confounder (U) and exposure (X). If separate bias
models for X and U are desired, use \code{adjust_uc_emc}.
}
\examples{
adjust_multinom_uc_emc(
 df_uc_emc,
 exposure = "Xstar",
 outcome = "Y",
 confounders = "C1",
 x1u0_model_coefs = c(-2.82, 1.62, 0.68, -0.06),
 x0u1_model_coefs = c(-0.20, 0.00, 0.68, -0.05),
 x1u1_model_coefs = c(-2.36, 1.62, 1.29, 0.27)
)

}
