% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_om.R
\name{adjust_omc}
\alias{adjust_omc}
\title{Adust for outcome misclassification.}
\usage{
adjust_omc(
  data,
  exposure,
  outcome,
  confounders = NULL,
  y_model_coefs,
  level = 0.95
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{adjust_omc()} was renamed to \code{adjust_om()}
}
\keyword{internal}
