% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_sel.R
\name{adjust_sel}
\alias{adjust_sel}
\title{Adust for selection bias.}
\usage{
adjust_sel(
  data,
  exposure,
  outcome,
  confounders = NULL,
  s_model_coefs,
  level = 0.95
)
}
\arguments{
\item{data}{Dataframe for analysis.}

\item{exposure}{String name of the exposure variable.}

\item{outcome}{String name of the outcome variable.}

\item{confounders}{String name(s) of the confounder(s).
A maximum of three confounders are allowed.}

\item{s_model_coefs}{The regression coefficients corresponding to the model:
\ifelse{html}{\out{logit(P(S=1)) = &beta;<sub>0</sub> +
&beta;<sub>1</sub>X + &beta;<sub>2</sub>Y, } where S represents
(binary) selection, X is the (binary) exposure, Y is the (binary)
outcome. The number of parameters is therefore 3.}{\eqn{logit(P(S=1)) =}}}

\item{level}{Value from 0-1 representing the full range of the confidence
interval. Default is 0.95.}
}
\description{
\code{adjust_sel} returns the exposure-outcome odds ratio and confidence
interval, adjusted for selection bias.
}
\details{
Details
}
\examples{
adjust_sel(
 evans,
 exposure = "SMK",
 outcome = "CHD",
 confounders = "AGE",
 s_model_coefs = c(0.05, 0.20, 0.20)
)

}
