% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_uc.R
\name{adjust_uc}
\alias{adjust_uc}
\title{Adust for uncontrolled confounding.}
\usage{
adjust_uc(
  data,
  exposure,
  outcome,
  confounders = NULL,
  u_model_coefs,
  level = 0.95
)
}
\arguments{
\item{data}{Dataframe for analysis.}

\item{exposure}{String name of the exposure variable.}

\item{outcome}{String name of the outcome variable.}

\item{confounders}{String name(s) of the confounder(s).
A maximum of three confounders are allowed.}

\item{u_model_coefs}{The regression coefficients corresponding to the model:
\ifelse{html}{\out{logit(P(U=1)) = &alpha;<sub>0</sub> +
&alpha;<sub>1</sub>X + &alpha;<sub>2</sub>Y +
&alpha;<sub>2+j</sub>C<sub>j</sub>, } where U is the (binary) unmeasured
confounder, X is the (binary) exposure, Y is the (binary) outcome,
C represents the vector of (binary) measured confounders (if any),
and j corresponds to the number of measured confounders.
The number of parameters therefore equals 3 + j.}{\eqn{logit(P(U=1)) =}}}

\item{level}{Value from 0-1 representing the full range of the confidence
interval. Default is 0.95.}
}
\description{
\code{adjust_uc} returns the exposure-outcome odds ratio and confidence
interval, adjusted for uncontrolled confounding.
}
\details{
description
}
\examples{
adjust_uc(
 evans,
 exposure = "SMK",
 outcome = "CHD",
 confounders = "AGE",
 u_model_coefs = c(0.10, 0.10, 0.10, 0.02),
)

}
