% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_uc_em_sel.R
\name{adjust_uc_emc_sel}
\alias{adjust_uc_emc_sel}
\title{Adust for uncontrolled confounding, exposure misclassification, and selection
bias.}
\usage{
adjust_uc_emc_sel(
  data_observed,
  u_model_coefs = NULL,
  x_model_coefs = NULL,
  x1u0_model_coefs = NULL,
  x0u1_model_coefs = NULL,
  x1u1_model_coefs = NULL,
  s_model_coefs,
  level = 0.95
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{adjust_uc_emc_sel()} was renamed to \code{adjust_uc_em_sel()}
}
\keyword{internal}
