% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uc_mc_sel.R
\name{adjust_uc_mc_sel}
\alias{adjust_uc_mc_sel}
\title{Adust for uncontrolled confounding, exposure misclassification, and selection bias.}
\usage{
adjust_uc_mc_sel(data, exposure, outcome, confounders = NULL,
  px1_u0_parameters, px0_u1_parameters, px1_u1_parameters, ps1_parameters,
  level = 0.95)
}
\arguments{
\item{data}{The data set.}

\item{exposure}{The variable corresponding to the exposure in the data.}

\item{outcome}{The variable corresponding to the outcome in the data.}

\item{confounders}{The variable(s) corresponding to the confounder(s) in the data.
A maximum of three confounders are allowed.}
}
\description{
\code{adjust_uc_mc_sel} returns the exposure-outcome odds ratio and confidence
interval, adjusted for uncontrolled confounding, exposure misclassificaiton,
and selection bias.
}
