% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uc_mc_sel2.R
\name{adjust_uc_mc_sel2}
\alias{adjust_uc_mc_sel2}
\title{Adust for uncontrolled confounding, exposure misclassification, and selection bias.}
\usage{
adjust_uc_mc_sel2(data, exposure, outcome, confounders = NULL,
  pu1_parameters, px1_parameters, ps1_parameters, level = 0.95)
}
\arguments{
\item{data}{The data set.}

\item{exposure}{The variable corresponding to the exposure in the data.}

\item{outcome}{The variable corresponding to the outcome in the data.}

\item{confounders}{The variable(s) corresponding to the confounder(s) in the data.
A maximum of three confounders are allowed.}

\item{pu1_parameters}{The regression coefficients corresponding to the model:
\ifelse{html}{\out{logit(P(U=1)) = &alpha;<sub>0</sub> + &alpha;<sub>1</sub>X + &alpha;<sub>2</sub>Y, } 
where U is the (binary) unmeasured confounder, X is the (binary) true exposure, Y is the (binary) 
outcome. The number of parameters therefore equals 3.}{\eqn{logit(P(U=1)) =}}}

\item{px1_parameters}{The regression coefficients corresponding to the model:
\ifelse{html}{\out{logit(P(X=1)) = &delta;<sub>0</sub> + &delta;<sub>1</sub>X* + &delta;<sub>2</sub>Y +
&delta;<sub>2+j</sub>C<sub>j</sub>, }where X represents (binary) true exposure, X* is the
(binary) misclassified exposure, Y is the (binary) outcome, C represents the vector of (binary)
measured confounders (if any), and j corresponds to the number of measured 
confounders. The number of parameters is therefore 3 + j.}{\eqn{logit(P(X=1)) =}}}

\item{ps1_parameters}{The regression coefficients corresponding to the model:
\ifelse{html}{\out{logit(P(S=1)) = &beta;<sub>0</sub> + &beta;<sub>1</sub>X* + &beta;<sub>2</sub>Y +
&beta;<sub>2+j</sub>C<sub>2+j</sub>, }where S represents (binary) selection, X* is the (binary) 
misclassified exposure, Y is the (binary) outcome, C represents the vector of (binary) 
measured confounders (if any), and j corresponds to the number of measured 
confounders. The number of parameters is therefore 3 + j.}{\eqn{logit(P(S=1)) =}}}

\item{level}{Number from 0-1 representing the range of the confidence interval. Default is .95.}
}
\description{
\code{adjust_uc_mc_sel2} returns the exposure-outcome odds ratio and confidence
interval, adjusted for uncontrolled confounding, exposure misclassificaiton,
and selection bias.
}
\details{
This function uses two separate logistic regression models to predict the uncontrolled 
confounder (U) and exposure (X). If a single bias model for jointly modeling X and U 
is desired use \code{adjust_uc_mc_sel}.
}
\examples{
adjust_uc_mc_sel2(df_uc_mc_sel, exposure = "Xstar", outcome = "Y", 
confounders = c("C1", "C2", "C3"), 
pu1_parameters = c(-.40, .38, .46), 
px1_parameters = c(-1.61, 2.71, .62, -.41, -.41, .40),
ps1_parameters = c(-.39, .40, .75, -.04, -.04, .05))

}
