% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_uc_om.R
\name{adjust_uc_omc}
\alias{adjust_uc_omc}
\title{Adust for uncontrolled confounding and outcome misclassification.}
\usage{
adjust_uc_omc(
  data,
  exposure,
  outcome,
  confounders = NULL,
  u_model_coefs = NULL,
  y_model_coefs = NULL,
  u0y1_model_coefs = NULL,
  u1y0_model_coefs = NULL,
  u1y1_model_coefs = NULL,
  level = 0.95
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{adjust_uc_omc_sel()} was renamed to \code{adjust_uc_om_sel()}
}
\keyword{internal}
