% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_uc_om_sel.R
\name{adjust_uc_omc_sel}
\alias{adjust_uc_omc_sel}
\title{Adust for uncontrolled confounding, outcome misclassification, and selection
bias.}
\usage{
adjust_uc_omc_sel(
  data_observed,
  u_model_coefs = NULL,
  y_model_coefs = NULL,
  u0y1_model_coefs = NULL,
  u1y0_model_coefs = NULL,
  u1y1_model_coefs = NULL,
  s_model_coefs,
  level = 0.95
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{adjust_uc_omc_sel()} was renamed to \code{adjust_uc_om_sel()}
}
\keyword{internal}
