% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_params.R
\name{bias_params}
\alias{bias_params}
\title{Represent bias parameters}
\usage{
bias_params(coef_list)
}
\arguments{
\item{coef_list}{List ...}
}
\description{
\code{bias_params} ...
}
\examples{
list_for_uc <- list(
  u = c(-0.19, 0.61, 0.70, -0.09, 0.10, -0.15)
)

bp <- bias_params(coef_list = list_for_uc)

}
