% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_data.R
\name{data_observed}
\alias{data_observed}
\title{Represent observed causal data}
\usage{
data_observed(data, exposure, outcome, confounders = NULL)
}
\arguments{
\item{data}{Dataframe for bias analysis.}

\item{exposure}{String name of the exposure variable.}

\item{outcome}{String name of the outcome variable.}

\item{confounders}{String name(s) of the confounder(s).}
}
\description{
\code{data_observed} combines the observed dataframe with specific identification
of the columns corresponding to the exposure, outcome, and confounders. It is
an essential input of all \code{adjust} functions.
}
\examples{
df <- data_observed(
  data = df_uc,
  exposure = "X",
  outcome = "Ystar",
  confounders = c("C1", "C2", "C3")
)

}
