% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_mc_sel.R
\docType{data}
\name{df_mc_sel}
\alias{df_mc_sel}
\title{Simulated dataset with exposure misclassification and selection bias}
\format{A dataframe with 100,000 rows and 5 variables:
\describe{
    \item{Xstar}{misclassified exposure, 1 = present and 0 = absent}
    \item{Y}{outcome, 1 = present and 0 = absent}
    \item{C1}{1st confounder, 1 = present and 0 = absent}
    \item{C2}{2nd confounder, 1 = present and 0 = absent}
    \item{C3}{3rd confounder, 1 = present and 0 = absent}
}}
\usage{
df_mc_sel
}
\description{
A dataset with two sources of bias, three known confounders, and 100,000 observations.
The true, unbiased exposure-outcome odds ratio = 2.
}
\keyword{datasets}
