% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_uc_emc_sel_source.R
\docType{data}
\name{df_uc_emc_sel_source}
\alias{df_uc_emc_sel_source}
\title{Data source for df_uc_emc_sel}
\format{
A dataframe with 100,000 rows and 8 variables:
\describe{
    \item{X}{true exposure, 1 = present and 0 = absent}
    \item{Y}{outcome, 1 = present and 0 = absent}
    \item{C1}{1st confounder, 1 = present and 0 = absent}
    \item{C2}{2nd confounder, 1 = present and 0 = absent}
    \item{C3}{3rd confounder, 1 = present and 0 = absent}
    \item{U}{unmeasured confounder, 1 = present and 0 = absent}
    \item{Xstar}{misclassified exposure, 1 = present and 0 = absent}
    \item{S}{selection, 1 = selected into the study and 0 = not selected into the study}
}
}
\usage{
df_uc_emc_sel_source
}
\description{
Data with complete information on the three sources of bias, three known
 confounders, and 100,000 observations. This data is used to derive
 df_uc_emc_sel and can be used to obtain bias parameters for purposes of
 validating the simultaneous multi-bias adjustment method with
 df_uc_emc_sel. Regressing Y on X, C1, C2, C3 and U shows that the true,
 unbiased exposure-outcome odds ratio = 2.
}
\keyword{datasets}
