% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evans.R
\docType{data}
\name{evans}
\alias{evans}
\title{Evans County dataset}
\format{A dataframe with 609 rows and 9 variables:
\describe{
    \item{ID}{subject identifiction}
    \item{CHD}{outcome variable; 1 = coronary heart disease}
    \item{AGE}{age (in years)}
    \item{CHL}{cholesterol, mg/dl}
    \item{SMK}{1 = subject has ever smoked}
    \item{ECG}{1 = presence of electrocardiogram abnormality}
    \item{DBP}{diastolic blood pressure, mmHg}
    \item{SBP}{systolic blood pressure, mmHg}
    \item{HPT}{1 = SBP greater than or equal to 160 or DBP greater than or equal to 95}
}}
\usage{
evans
}
\description{
Data from cohort study in which white males in Evans County were followed for 7 years, 
with coronary heart disease as the outcome of interest.
}
\keyword{datasets}
