% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multibias-package.R
\docType{package}
\name{multibias-package}
\alias{multibias}
\alias{multibias-package}
\title{multibias: Simultaneous Multi-Bias Adjustment}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Quantify the causal effect of a binary exposure on a binary outcome with adjustment for multiple biases. The functions can simultaneously adjust for any combination of uncontrolled confounding, exposure/outcome misclassification, and selection bias. The underlying method generalizes the concept of combining inverse probability of selection weighting with predictive value weighting. Simultaneous multi-bias analysis can be used to enhance the validity and transparency of real-world evidence obtained from observational, longitudinal studies. Based on the work from Paul Brendel, Aracelis Torres, and Onyebuchi Arah (2023) \doi{10.1093/ije/dyad001}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pcbrendel/multibias}
  \item Report bugs at \url{https://github.com/pcbrendel/multibias/issues}
}

}
\author{
\strong{Maintainer}: Paul Brendel \email{pcbrendel@gmail.com} [copyright holder]

}
\keyword{internal}
