% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_em.R
\name{adjust_emc}
\alias{adjust_emc}
\title{Adust for exposure misclassification.}
\usage{
adjust_emc(
  data,
  exposure,
  outcome,
  confounders = NULL,
  x_model_coefs,
  level = 0.95
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{adjust_emc()} was renamed to \code{adjust_em()}
}
\keyword{internal}
