% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adust_om_sel.R
\name{adjust_omc_sel}
\alias{adjust_omc_sel}
\title{Adust for outcome misclassification and selection bias.}
\usage{
adjust_omc_sel(
  data,
  exposure,
  outcome,
  confounders = NULL,
  y_model_coefs,
  s_model_coefs,
  level = 0.95
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{adjust_omc_sel()} was renamed to \code{adjust_om_sel()}
}
\keyword{internal}
