% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_uc_emc.R
\name{adjust_uc_emc}
\alias{adjust_uc_emc}
\title{Adust for uncontrolled confounding and exposure misclassification.}
\usage{
adjust_uc_emc(
  data,
  exposure,
  outcome,
  confounders = NULL,
  pu1_parameters,
  px1_parameters,
  level = 0.95
)
}
\arguments{
\item{data}{The data set.}

\item{exposure}{The name of the exposure variable in the data.}

\item{outcome}{The name of the outcome variable in the data.}

\item{confounders}{The variable name(s) of the confounder(s) in the data.
A maximum of three confounders are allowed.}

\item{pu1_parameters}{The regression coefficients corresponding to the model:
\ifelse{html}{\out{logit(P(U=1)) = &alpha;<sub>0</sub> +
&alpha;<sub>1</sub>X + &alpha;<sub>2</sub>Y, }
where U is the (binary) unmeasured confounder, X is the (binary) true
exposure, Y is the (binary) outcome. The number of parameters therefore
equals 3.}{\eqn{logit(P(U=1)) =}}}

\item{px1_parameters}{The regression coefficients corresponding to the model:
\ifelse{html}{\out{logit(P(X=1)) = &delta;<sub>0</sub> +
&delta;<sub>1</sub>X* + &delta;<sub>2</sub>Y +
&delta;<sub>2+j</sub>C<sub>j</sub>, } where X represents (binary) true
exposure, X* is the (binary) misclassified exposure, Y is the (binary)
outcome, C represents the vector of (binary) measured confounders (if any),
and j corresponds to the number of measured confounders. The number of
parameters is therefore 3 + j.}{\eqn{logit(P(X=1)) =}}}

\item{level}{Numeric value from 0-1 representing the range of the confidence
interval. The default value is 0.95.}
}
\description{
\code{adjust_uc_emc} returns the exposure-outcome odds ratio and confidence
interval, adjusted for uncontrolled confounding and exposure
misclassificaiton.
}
\details{
This function uses two separate logistic regression models to predict the
uncontrolled confounder (U) and exposure (X). If a single bias model for
jointly modeling X and U is desired use \code{adjust_multinom_uc_emc}.
}
\examples{
adjust_uc_emc(
 df_uc_mc,
 exposure = "Xstar",
 outcome = "Y",
 confounders = c("C1", "C2", "C3"),
 pu1_parameters = c(-.40, .38, .46),
 px1_parameters = c(-1.21, 1.69, .76, -.41, -.41, .41)
)

}
