% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_uc_emc_sel.R
\name{adjust_uc_emc_sel}
\alias{adjust_uc_emc_sel}
\title{Adust for uncontrolled confounding, exposure misclassification, and selection
bias.}
\usage{
adjust_uc_emc_sel(
  data,
  exposure,
  outcome,
  confounders = NULL,
  u_model_coefs,
  x_model_coefs,
  s_model_coefs,
  level = 0.95
)
}
\arguments{
\item{data}{Dataframe for analysis.}

\item{exposure}{String name of the exposure variable.}

\item{outcome}{String name of the outcome variable.}

\item{confounders}{String name(s) of the confounder(s).
A maximum of three confounders are allowed.}

\item{u_model_coefs}{The regression coefficients corresponding to the model:
\ifelse{html}{\out{logit(P(U=1)) = &alpha;<sub>0</sub> +
&alpha;<sub>1</sub>X + &alpha;<sub>2</sub>Y, } where U is the (binary)
unmeasured confounder, X is the (binary) true exposure, Y is the (binary)
outcome. The number of parameters therefore
equals 3.}{\eqn{logit(P(U=1)) =}}}

\item{x_model_coefs}{The regression coefficients corresponding to the model:
\ifelse{html}{\out{logit(P(X=1)) = &delta;<sub>0</sub> +
&delta;<sub>1</sub>X* + &delta;<sub>2</sub>Y +
&delta;<sub>2+j</sub>C<sub>j</sub>, } where X represents (binary) true
exposure, X* is the (binary) misclassified exposure, Y is the
(binary) outcome, C represents the vector of (binary)
measured confounders (if any), and j corresponds to the number of measured
confounders. The number of parameters is therefore
3 + j.}{\eqn{logit(P(X=1)) =}}}

\item{s_model_coefs}{The regression coefficients corresponding to the model:
\ifelse{html}{\out{logit(P(S=1)) = &beta;<sub>0</sub> +
&beta;<sub>1</sub>X* + &beta;<sub>2</sub>Y +
&beta;<sub>2+j</sub>C<sub>2+j</sub>, } where S represents (binary)
selection, X* is the (binary) misclassified exposure, Y is the (binary)
outcome, C represents the vector of (binary) measured confounders (if any),
and j corresponds to the number of measured confounders.
The number of parameters is therefore 3 + j.}{\eqn{logit(P(S=1)) =}}}

\item{level}{Value from 0-1 representing the full range of the confidence
interval. Default is 0.95.}
}
\description{
\code{adjust_uc_emc_sel} returns the exposure-outcome odds ratio and
confidence interval, adjusted for uncontrolled confounding, exposure
misclassificaiton, and selection bias.
}
\details{
This function uses two separate logistic regression models to predict the
uncontrolled confounder (U) and exposure (X). If a single bias model for
jointly modeling X and U is desired use \code{adjust_uc_emc_sel}.
}
\examples{
adjust_uc_emc_sel(
 df_uc_emc_sel,
 exposure = "Xstar",
 outcome = "Y",
 confounders = c("C1", "C2", "C3"),
 u_model_coefs = c(-0.32, 0.59, 0.69),
 x_model_coefs = c(-2.44, 1.62, 0.72, 0.32, -0.15, 0.85),
 s_model_coefs = c(0.00, 0.26, 0.78, 0.03, -0.02, 0.10)
)

}
