% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uc_mc.R
\name{adjust_uc_mc}
\alias{adjust_uc_mc}
\title{Adust for uncontrolled confounding and exposure misclassification.}
\usage{
adjust_uc_mc(data, exposure, outcome, confounders = NULL,
  px1_u0_parameters, px0_u1_parameters, px1_u1_parameters, level = 0.95)
}
\arguments{
\item{data}{The data set.}

\item{exposure}{The variable corresponding to the exposure in the data.}

\item{outcome}{The variable corresponding to the outcome in the data.}

\item{confounders}{The variable(s) corresponding to the confounder(s) in the data.
A maximum of three confounders are allowed.}

\item{px1_u0_parameters}{The regression coefficients corresponding to the model: .}

\item{px0_u1_parameters}{The regression coefficients corresponding to the model: .}

\item{px1_u1_parameters}{The regression coefficients corresponding to the model: .}
}
\value{
The exposure-outcome odds ratio, adjusted for uncontrolled
  confounding and exposure misclassification, and its confidence interval.
}
\description{
\code{adjust_uc_mc} returns the exposure-outcome odds ratio and confidence
interval, adjusted for uncontrolled confounding and exposure
misclassificaiton.
}
\details{
Details
}
\author{
Paul Brendel
}
