% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uc_mc_sel.R
\name{adjust_uc_mc_sel}
\alias{adjust_uc_mc_sel}
\title{Adust for uncontrolled confounding, exposure misclassification, and selection bias.}
\usage{
adjust_uc_mc_sel(data, exposure, outcome, confounders = NULL,
  px1_u0_parameters, px0_u1_parameters, px1_u1_parameters, ps1_parameters,
  level = 0.95)
}
\arguments{
\item{data}{The data set.}

\item{exposure}{The variable corresponding to the exposure in the data.}

\item{outcome}{The variable corresponding to the outcome in the data.}

\item{confounders}{The variable(s) corresponding to the confounder(s) in the data.
A maximum of three confounders are allowed.}

\item{px1_u0_parameters}{The regression coefficients corresponding to the model:
\ifelse{html}{\out{log(P(X=1,U=0)/P(X=0,U=0)) = &gamma;<sub>1,0</sub> + &gamma;<sub>1,1</sub>X* + 
&gamma;<sub>1,2</sub>Y + &gamma;<sub>1,2+j</sub>C<sub>2+j</sub>, }where X is the true (binary) exposure,
U is the (binary) unmeasured confounder, X* is the (binary) misclassified exposure, Y is the (binary)
outcome, C represents the vector of (binary) measured confounders (if any), and j corresponds to the 
number of measured confounders.}{\eqn{log(P(X=1,U=0)/P(X=0,U=0)) =}}}

\item{px0_u1_parameters}{The regression coefficients corresponding to the model:
\ifelse{html}{\out{log(P(X=0,U=1)/P(X=0,U=0)) = &gamma;<sub>2,0</sub> + &gamma;<sub>2,1</sub>X* + 
&gamma;<sub>2,2</sub>Y + &gamma;<sub>2,2+j</sub>C<sub>2+j</sub>, }where X is the true (binary) exposure,
U is the (binary) unmeasured confounder, X* is the (binary) misclassified exposure, Y is the (binary)
outcome, C represents the vector of (binary) measured confounders (if any), and j corresponds to the
number of measured confounders.}{\eqn{log(P(X=0,U=1)/P(X=0,U=0)) =}}}

\item{px1_u1_parameters}{The regression coefficients corresponding to the model: 
\ifelse{html}{\out{log(P(X=1,U=1)/P(X=0,U=0)) = &gamma;<sub>3,0</sub> + &gamma;<sub>3,1</sub>X* + 
&gamma;<sub>3,2</sub>Y + &gamma;<sub>3,2+j</sub>C<sub>2+j</sub>, }where X is the true (binary) exposure,
U is the (binary) unmeasured confounder, X* is the (binary) misclassified exposure, Y is the (binary)
outcome, C represents the vector of (binary) measured confounders (if any), and j corresponds to the
number of measured confounders.}{\eqn{log(P(X=1,U=1)/P(X=0,U=0)) =}}}

\item{ps1_parameters}{The regression coefficients corresponding to the model:
\ifelse{html}{\out{logit(P(S=1)) = &beta;<sub>0</sub> + &beta;<sub>1</sub>X* + &beta;<sub>2</sub>Y +
&beta;<sub>2+j</sub>C<sub>2+j</sub>, }where S represents (binary) selection, X* is the (binary) 
misclassified exposure, Y is the (binary) outcome, C represents the vector of (binary) 
measured confounders (if any), and j corresponds to the number of measured 
confounders.}{\eqn{logit(P(S=1)) =}}}

\item{level}{Number from 0-1 representing the range of the confidence interval. Default is .95.}
}
\description{
\code{adjust_uc_mc_sel} returns the exposure-outcome odds ratio and confidence
interval, adjusted for uncontrolled confounding, exposure misclassificaiton,
and selection bias.
}
\details{
This function uses one bias model, a multinomial logistic regression model, to predict 
the uncontrolled confounder (U) and exposure (X). If separate bias models for X and U 
are desired, use \code{adjust_uc_mc_2}.
}
\examples{

}
