% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_uc_sel.R
\name{adjust_uc_sel}
\alias{adjust_uc_sel}
\title{Adust for uncontrolled confounding and selection bias.}
\usage{
adjust_uc_sel(
  data,
  exposure,
  outcome,
  confounders = NULL,
  u_model_coefs,
  s_model_coefs,
  level = 0.95
)
}
\arguments{
\item{data}{Dataframe for analysis.}

\item{exposure}{String name of the exposure variable.}

\item{outcome}{String name of the outcome variable.}

\item{confounders}{String name(s) of the confounder(s).
A maximum of three confounders are allowed.}

\item{u_model_coefs}{The regression coefficients corresponding to the model:
\ifelse{html}{\out{logit(P(U=1)) = &alpha;<sub>0</sub> +
&alpha;<sub>1</sub>X + &alpha;<sub>2</sub>Y +
&alpha;<sub>2+j</sub>C<sub>j</sub>, } where U is the (binary) unmeasured
confounder, X is the (binary) exposure, Y is the (binary) outcome, C
represents the vector of (binary) measured confounders (if any), and j
corresponds to the number of measured confounders. The number of parameters
therefore equals 3 + j.}{\eqn{logit(P(U=1)) =}}}

\item{s_model_coefs}{The regression coefficients corresponding to the model:
\ifelse{html}{\out{logit(P(S=1)) = &beta;<sub>0</sub> +
&beta;<sub>1</sub>X + &beta;<sub>2</sub>Y,} where S represents (binary)
selection, X is the (binary) exposure, and Y is the (binary) outcome.
The number of parameters therefore equals 3.}{\eqn{logit(P(S=1)) =}}}

\item{level}{Value from 0-1 representing the full range of the confidence
interval. Default is 0.95.}
}
\description{
\code{adjust_uc_sel} returns the exposure-outcome odds ratio and confidence
interval, adjusted for uncontrolled confounding and exposure
misclassificaiton.
}
\details{
Values for the regression coefficients can be applied as
fixed values or as single draws from a probability
distribution (ex: \code{rnorm(1, mean = 2, sd = 1)}). The latter has
the advantage of allowing the researcher to capture the uncertainty
in the bias parameter estimates. To incorporate this uncertainty in the
estimate and confidence interval, this function should be run in loop across
bootstrap samples of the dataframe for analysis. The estimate and
confidence interval would then be obtained from the median and quantiles
of the distribution of odds ratio estimates.
}
\examples{
adjust_uc_sel(
  df_uc_sel,
  exposure = "X",
  outcome = "Y",
  confounders = "C1",
  u_model_coefs = c(-0.36, 0.96, 1.11, -0.12),
  s_model_coefs = c(0.02, 0.89, 0.90)
)

}
