% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_uc_emc.R
\docType{data}
\name{df_uc_emc}
\alias{df_uc_emc}
\title{Simulated data with uncontrolled confounding and exposure misclassification}
\format{
A dataframe with 100,000 rows and 3 columns:
\describe{
    \item{Xstar}{misclassified exposure, 1 = present and 0 = absent}
    \item{Y}{outcome, 1 = present and 0 = absent}
    \item{C1}{confounder, 1 = present and 0 = absent}
}
}
\usage{
df_uc_emc
}
\description{
Data containing two sources of bias, a known confounder, and
 100,000 observations. This data is obtained from \code{df_uc_emc_source}
 by removing the columns X and U. The resulting data corresponds to
 what a researcher would see in the real-world: a misclassified exposure,
 Xstar, and missing data on a confounder U. As seen in
 \code{df_uc_emc_source}, the true, unbiased exposure-outcome odds ratio = 2.
}
\keyword{datasets}
